projeto_sptrans/data_processor/requirements.txt
aws-hadoop==1.0.dev2
bcrypt==4.2.0
boto3==1.35.24
botocore==1.35.24
cffi==1.17.1
coverage==7.6.1
cryptography==43.0.1
decorator==5.1.1
Deprecated==1.2.14
fabric==3.2.2
gitdb==4.0.11
GitPython==0.3.6
invoke==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
paramiko==3.5.0
pep8==1.7.1
py4j==0.10.9.7
pycparser==2.22
PyNaCl==1.5.0
pypi-publisher==0.0.4
pyspark==3.5.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-terraform==0.10.1
s3transfer==0.10.2
six==1.16.0
smmap==5.0.1
urllib3==2.2.3
wrapt==1.16.0